packages/pdfsys-cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdfsys-cli"
version = "0.0.1"
description = "Unified CLI for pdfsys pipeline — run individual stages or chain them via YAML config."
requires-python = ">=3.11"
dependencies = [
    "pdfsys-core",
    "pdfsys-router",
    "pdfsys-parser-mupdf",
    "pdfsys-layout-analyser",
    "pdfsys-parser-pipeline",
    "pdfsys-parser-vlm",
    "pdfsys-bench",
    "pyyaml>=6.0",
]

[project.scripts]
pdfsys = "pdfsys_cli.__main__:main"

[tool.uv.sources]
pdfsys-core             = { workspace = true }
pdfsys-router           = { workspace = true }
pdfsys-parser-mupdf     = { workspace = true }
pdfsys-layout-analyser  = { workspace = true }
pdfsys-parser-pipeline  = { workspace = true }
pdfsys-parser-vlm       = { workspace = true }
pdfsys-bench            = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/pdfsys_cli"]
